add & update tests
