+++ tests/test_quiz_constants.py
@@ -2,7 +2,7 @@ import os
 from orm_study.quiz.constants import CHAPTER, CHAPTER_CHOICE
 
 
-CHAPTER_DIRS = os.listdir("orm_study/quiz/python")
+CHAPTER_DIRS = os.listdir("orm_study/quiz/_python")
 
 
 def test_check_chapters_all_set() -> None:
